Refactor KeyReplacer folder
