+++ src/azure-cli/azure/cli/command_modules/iot/tests/hybrid_2020_09_01/test_iot_commands.py
@@ -6,7 +6,7 @@
 
 from azure.cli.testsdk import ResourceGroupPreparer, ScenarioTest, StorageAccountPreparer
 from azure.cli.testsdk.scenario_tests import AllowLargeResponse
-from ..utils.recording_processors import KeyReplacer
+from ..utils import KeyReplacer
 import unittest
