+++ src/azure-cli/azure/cli/command_modules/iot/tests/latest/test_iot_commands.py
@@ -12,7 +12,7 @@ from azure.cli.testsdk.scenario_tests import AllowLargeResponse
 from azure.mgmt.iothub.models import RoutingSource
 from azure.cli.command_modules.iot.shared import IdentityType
 from azure.core.exceptions import HttpResponseError
-from ..utils.recording_processors import KeyReplacer
+from ..utils import KeyReplacer
 
 
 class IoTHubTest(ScenarioTest):
