+++ src/azure-cli/azure/cli/command_modules/iot/tests/utils.py
