--- src/azure-cli/azure/cli/command_modules/iot/tests/utils/__init__.py
@@ -1,4 +1,0 @@
-# --------------------------------------------------------------------------------------------
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License. See License.txt in the project root for license information.
-# --------------------------------------------------------------------------------------------
